IN-224 Added posibility to specify time slots for schedule-holiday mappings.

+++ src/schedule-holiday/schedule-holiday.service.ts
@@ -11,6 +11,10 @@ export class ScheduleHolidayService {
   ) {}
 
   async removeAll(scheduleHolidays: ScheduleHoliday[]) {
+    if (!scheduleHolidays || !scheduleHolidays.length) {
+      return [];
+    }
+
     return await this.scheduleHolidayRepository.remove(scheduleHolidays);
   }
 }
